rename api-configuration service
